split out battle dialogs
